Technically floor() returns a float so we should cast this to an int before using it as an array key

+++ classes/schen/Utils/Uploader.php
@@ -44,7 +44,7 @@ class Uploader
     public function bytesToSize1024($bytes, $precision = 2)
     {
         $unit = array('B', 'KB', 'MB');
-        return @round($bytes / pow(1024, ($i = floor(log($bytes, 1024)))), $precision) . ' ' . $unit[$i];
+        return @round($bytes / pow(1024, ($i = (int)floor(log($bytes, 1024)))), $precision) . ' ' . $unit[$i];
     }
 
     public function upload()
